requirements.txt
pandas
scipy
numpy
scikit-learn
joblib
flask
nltk
seaborn
wordcloud  
bert-extractive-summarizer
transformers
Jinja2
torch